+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to s3sup are documented in this file with each release
 version. ChangeLog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2019-06-10
+### Changed
+ - Prevent HTML files referencing static assets (stylesheets/scripts/images)
+   before they exist on S3. Achieved by applying a crude but effective ordering
+   to the upload sequence to ensure HTML files are uploaded after all other
+   assets.
+
+
 ## [0.4.0] - 2019-05-03
 ### Changed
  - Remote catalogue now a gzipped SQLite database rather than an uncompressed
@@ -10,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    automatic and requires no intervention. One caveat is that after running an
    `s3sup push` with this version or newer, it will no longer be possible to
    use _older_ versions of s3sup (<= 0.3.0) with the same project.
+
 
 ## [0.3.0] - 2019-04-09
 ### Added